split makefile scripts
