+++ auto/thirdparty.lua
@@ -1,0 +1,57 @@
+--------------------------------------------------------------
+-- All SDK configs
+CommercialSDKConfigs = {}
+function DefineSDKConfig(sdkDirectoryMacro)
+	local sdkDirectory = os.getenv(sdkDirectoryMacro)
+	if sdkDirectory and not os.isdir(sdkDirectory) then
+		sdkDirectory = nil
+	end
+
+	if sdkDirectory then
+		print(sdkDirectoryMacro.." found...")
+		local sdkConfig = {}
+		sdkConfig.include = path.join(sdkDirectory, "include")
+		sdkConfig.lib_dir = path.join(sdkDirectory, "lib")
+		allLinkLibs = os.matchfiles(sdkConfig.lib_dir.."/*.lib")
+		sdkConfig.lib_names = { table.unpack(allLinkLibs) }
+		sdkConfig.dll_paths = { path.join(sdkDirectory, "runtime") }
+		CommercialSDKConfigs[sdkDirectoryMacro] = sdkConfig
+	else
+		print(sdkDirectoryMacro.." not found, skip related features...")
+	end
+end
+
+function CheckSDKExists(sdkDirectoryMacro)
+	return CommercialSDKConfigs[sdkDirectoryMacro] ~= nil
+end
+
+DefineSDKConfig("FBX_SDK_DIR")
+DefineSDKConfig("PHYSX_SDK_DIR")
+DefineSDKConfig("SPEEDTREE_SDK_DIR")
+--------------------------------------------------------------
+
+-- Define ThirdParty projects
+function DeclareExternalProject(projectName, projectKind, projectPath)
+	-- Same with projectName by default
+	projectPath = projectPath or projectName
+
+	externalproject(projectName)
+		kind(projectKind)
+		location(projectPath)
+end
+
+group "ThirdParty/assimp"
+	DeclareExternalProject("assimp", "StaticLib", path.join(RootPath, "build/assimp/code"))
+		dependson { "zlibstatic" }
+	DeclareExternalProject("zlibstatic", "StaticLib", path.join(RootPath, "build/assimp/contrib/zlib"))
+group ""
+
+function SetupAssimpLib(assimpLibName, configName)
+	libdirs {
+		path.join(RootPath, "build/assimp/lib", configName),
+	}
+
+	links {
+		assimpLibName,
+	}
+end
